Move query_scripts->query_modules, python_scripts, dataproc_scripts to root level. Dataproc scripts to import from query_modules
