+++ dataproc_scripts/annotate_dataset.py
@@ -1,0 +1,69 @@
+#!/usr/bin/env python3
+
+"""
+Dataproc script to annotate cohort.
+"""
+
+import click
+import coloredlogs
+import hail as hl
+
+from cpg_utils import to_path
+from cpg_utils.config import get_config
+
+from query_modules.seqr_loader import subset_mt_to_samples, annotate_dataset_mt
+
+fmt = '%(asctime)s %(levelname)s (%(name)s %(lineno)s): %(message)s'
+coloredlogs.install(level='DEBUG', fmt=fmt)
+
+
+@click.command()
+@click.option(
+    '--mt-path',
+    'mt_path',
+    required=True,
+)
+@click.option(
+    '--sample-ids',
+    'sample_ids_path',
+    required=True,
+)
+@click.option(
+    '--out-mt-path',
+    'out_mt_path',
+    required=True,
+)
+@click.option(
+    '--checkpoint-prefix',
+    'checkpoint_prefix',
+    required=True,
+)
+def main(
+    mt_path: str,
+    sample_ids_path: str,
+    out_mt_path: str,
+    checkpoint_prefix: str,
+):
+    hl.init(default_reference='GRCh38')
+
+    with to_path(sample_ids_path).open() as f:
+        sample_ids = f.read().strip().split(',')
+
+    subset_mt_path = to_path(checkpoint_prefix) / 'cohort-subset.mt'
+
+    subset_mt_to_samples(
+        mt_path=mt_path,
+        sample_ids=sample_ids,
+        out_mt_path=str(subset_mt_path),
+    )
+
+    annotate_dataset_mt(
+        mt_path=str(subset_mt_path),
+        out_mt_path=out_mt_path,
+        checkpoint_prefix=checkpoint_prefix,
+        overwrite=not get_config()['workflow'].get('check_intermediates'),
+    )
+
+
+if __name__ == '__main__':
+    main()  # pylint: disable=E1120
