--- jobs/dataproc_scripts/mt_to_es.py
@@ -1,605 +1,0 @@
-#!/usr/bin/env python3
-
-"""
-Hail script to submit on a dataproc cluster. 
-
-Loads the matrix table into an ElasticSearch index.
-"""
-
-import logging
-import math
-from pprint import pformat
-
-import click
-import hail as hl
-import coloredlogs
-import elasticsearch
-
-from cpg_utils import to_path
-from cpg_utils.cloud import read_secret
-from cpg_utils.config import get_config
-
-from hail_scripts.elasticsearch.hail_elasticsearch_client import HailElasticsearchClient
-
-
-fmt = '%(asctime)s %(levelname)s (%(name)s %(lineno)s): %(message)s'
-coloredlogs.install(level='DEBUG', fmt=fmt)
-
-
-class HailElasticsearchClientV8(HailElasticsearchClient):
-    """
-    The Broad's seqr-loading-pipelines pins the Elasticsearch client to v7. We use v8, 
-    so we are overriding the class to adjust for Elasticsearch v8.
-    """
-
-    # noinspection PyMissingConstructor
-    def __init__(
-        self,
-        host: str,
-        port: str,
-        es_username: str,
-        es_password: str,
-    ):
-        """
-        Overriding base __init__: the difference with v7 is that in v8, 
-        elasticsearch.Elasticsearch constructor takes one URL string, in contrast
-        with v7 which takes "host" and "port" strings separately. Note that we are
-        not calling the base init, which would fail on v8. Instead, we are completely
-        overriding init.
-        """
-        self._host = host
-        self._port = port
-        self._es_username = es_username
-        self._es_password = es_password
-        self._es_use_ssl = True
-
-        auth = (self._es_username, self._es_password) if self._es_password else None
-
-        if not host.startswith('http://') or not host.startswith('https://'):
-            scheme = 'https' if self._es_use_ssl else 'http'
-            host = f'{scheme}://{host}'
-        _host = f'{host}:{port}'
-        self.es = elasticsearch.Elasticsearch(_host, basic_auth=auth)
-
-        # check connection
-        logging.info(pformat(self.es.info()))
-        
-    def export_table_to_elasticsearch(self, *args, **kwargs):
-        """Override to adjust for ES V7."""
-
-        # Copied from older hail_scripts/v02/utils/elasticsearch_client.py:
-        # https://github.com/populationgenomics/hail-elasticsearch-pipelines/blob/main/hail_scripts/v02/utils/elasticsearch_client.py#L128-L133
-        # that's before elasticsearch in upstream seqr-loading-pipelines was pinned to v7.
-        # Without this config, ES API errors with the following:
-        # > Cannot detect ES version - typically this happens if the network/Elasticsearch
-        # cluster is not accessible or when targeting a WAN/Cloud instance without the 
-        # proper setting 'es.nodes.wan.only'
-        if self._es_use_ssl:
-            kwargs.setdefault('elasticsearch_config', {}).update({
-                'es.net.ssl': 'true',
-                # If using SSL, the instance is likely managed, in which case we
-                # can't discover nodes.
-                'es.nodes.wan.only': 'true',
-            })
-        
-        # deprecated in ES=V8:
-        kwargs['index_type_name'] = ''
-        
-        super().export_table_to_elasticsearch(*args, **kwargs)
-
-
-@click.command()
-@click.option(
-    '--mt-path',
-    'mt_path',
-    required=True,
-)
-@click.option(
-    '--es-index',
-    'es_index',
-    type=click.STRING,
-    help='Elasticsearch index. Usually the dataset name. Will be lowercased',
-    required=True,
-)
-@click.option(
-    '--done-flag-path',
-    'done_flag_path',
-    help='File to touch in the end',
-)
-@click.option(
-    '--liftover-path',
-    'liftover_path',
-    help='Path to liftover chain',
-    required=True,
-)
-@click.option(
-    '--es-password',
-    'password',
-)
-def main(
-    mt_path: str,
-    es_index: str,
-    done_flag_path: str,
-    liftover_path: str,
-    password: str = None,
-):
-    """
-    Entry point.
-    """
-    es_index = es_index.lower()
-    hl.init(default_reference='GRCh38')
-
-    host = get_config()['elasticsearch']['host']
-    port = str(get_config()['elasticsearch']['port'])
-    username = get_config()['elasticsearch']['username']
-    project_id = get_config()['elasticsearch']['password_project_id']
-    secret_name = get_config()['elasticsearch']['password_secret_id']
-    password = password or read_secret(
-        project_id=project_id,
-        secret_name=secret_name,
-        fail_gracefully=False,
-    )
-    assert password
-
-    print(
-        f'Connecting to ElasticSearch: host="{host}", port="{port}", user="{username}"'
-    )
-    print(f'Reading passport from secret "{secret_name}" in project "{project_id}"')
-    es = HailElasticsearchClientV8(
-        host=host,
-        port=port,
-        es_username=username,
-        es_password=password,
-    )
-
-    mt = hl.read_matrix_table(mt_path)
-    # Annotate GRCh37 coordinates here, until liftover is supported by Batch Backend
-    # and can be moved to annotate_cohort.
-    logging.info('Adding GRCh37 coords')
-    rg37 = hl.get_reference('GRCh37')
-    rg38 = hl.get_reference('GRCh38')
-    rg38.add_liftover(liftover_path, rg37)
-    mt = mt.annotate_rows(rg37_locus=hl.liftover(mt.locus, 'GRCh37'))
-    mt = mt.annotate_rows(info=mt.info.drop('InbreedingCoeff'))
-
-    logging.info('Getting rows and exporting to the ES')
-    row_ht = elasticsearch_row(mt)
-    es_shards = _mt_num_shards(mt)
-
-    es.export_table_to_elasticsearch(
-        row_ht,
-        index_name=es_index,
-        num_shards=es_shards,
-        write_null_values=True,
-    )
-    _cleanup(es, es_index, es_shards)
-    with to_path(done_flag_path).open('w') as f:
-        f.write('done')
-
-
-def elasticsearch_row(mt: hl.MatrixTable):
-    """
-    Prepares the mt for export.
-    - Flattens nested structs
-    - drops locus and alleles key
-    Borrowed from:
-    https://github.com/broadinstitute/hail-elasticsearch-pipelines/blob/495f0d1b4d49542557ca5cccf98a23fc627260bf/luigi_pipeline/lib/model/seqr_mt_schema.py
-    """
-    # Converts a mt to the row equivalent.
-    ht = mt.rows()
-    # Converts nested structs into one field, e.g. {a: {b: 1}} => a.b: 1
-    ht = ht.drop('vep').flatten()
-    # When flattening, the table is unkeyed, which causes problems because our locus
-    # and alleles should not be normal fields. We can also re-key, but I believe this
-    # is computational?
-    ht = ht.drop(ht.locus, ht.alleles)
-    ht.describe()
-    return ht
-
-
-def _mt_num_shards(mt):
-    """
-    Calculate the number of shards from the number of variants and samples.
-    """
-    denominator = 1.4 * 10**9
-    calculated_num_shards = math.ceil((mt.count_rows() * mt.count_cols()) / denominator)
-    return calculated_num_shards
-
-
-def _cleanup(es, es_index, es_shards):
-    # Current disk configuration requires the previous index to be deleted prior to large indices, ~1TB, transferring off loading nodes
-    if es_shards < 25:
-        es.wait_for_shard_transfer(es_index)
-
-
-# LOADING_NODES_NAME = 'elasticsearch-es-data-loading*'
-# 
-# 
-# def struct_to_dict(struct) -> dict:
-#     """Convert Hail query struct to dict"""
-#     return {
-#         k: dict(struct_to_dict(v)) if isinstance(v, hl.utils.Struct) else v
-#         for k, v in struct.items()
-#     }
-# 
-# 
-# class HailElasticsearchClient:
-#     """
-#     Wrapper around Elasticsearch client that loads matrix table
-#     """
-# 
-#     def __init__(
-#         self,
-#         host,
-#         port,
-#         es_username,
-#         es_password,
-#         es_use_ssl=False,
-#     ):
-#         """Constructor.
-# 
-#         Args:
-#             host (str): Elasticsearch server host
-#             port (str): Elasticsearch server port
-#             es_username (str): Elasticsearch username
-#             es_password (str): Elasticsearch password
-#             es_use_ssl (bool): Whether to use SSL for ElasticSearch connections
-#         """
-# 
-#         self._host = host
-#         self._port = port
-#         self._es_username = es_username
-#         self._es_password = es_password
-#         self._es_use_ssl = es_use_ssl
-# 
-#         auth = (self._es_username, self._es_password) if self._es_password else None
-# 
-#         if not host.startswith('http://') or not host.startswith('https://'):
-#             scheme = 'https' if es_use_ssl else 'http'
-#             host = f'{scheme}://{host}'
-#         _host = f'{host}:{port}'
-#         self.es = elasticsearch.Elasticsearch(_host, basic_auth=auth)
-# 
-#         # check connection
-#         logging.info(pformat(self.es.info()))
-# 
-#     def create_index(self, index_name, elasticsearch_schema, num_shards=1, _meta=None):
-#         """Calls es.indices.create to create an elasticsearch index with the appropriate mapping.
-# 
-#         Args:
-#             index_name (str): elasticsearch index mapping
-#             elasticsearch_schema (dict): elasticsearch mapping 'properties' dictionary
-#             num_shards (int): how many shards the index will contain
-#             _meta (dict): optional _meta info for this index
-#                 (see https://www.elastic.co/guide/en/elasticsearch/reference/current/mapping-meta-field.html)
-#         """
-# 
-#         self.create_or_update_mapping(
-#             index_name,
-#             elasticsearch_schema,
-#             num_shards=num_shards,
-#             _meta=_meta,
-#             create_only=True,
-#         )
-# 
-#     def create_or_update_mapping(
-#         self,
-#         index_name,
-#         elasticsearch_schema,
-#         num_shards=1,
-#         _meta=None,
-#         create_only=False,
-#     ):
-#         """Calls es.indices.create or es.indices.put_mapping to create or update an elasticsearch index mapping.
-# 
-#         Args:
-#             index_name (str): elasticsearch index mapping
-#             elasticsearch_schema (dict): elasticsearch mapping 'properties' dictionary
-#             num_shards (int): how many shards the index will contain
-#             _meta (dict): optional _meta info for this index
-#                 (see https://www.elastic.co/guide/en/elasticsearch/reference/current/mapping-meta-field.html)
-#             create_only (bool): only allow index creation, throws an error if index already exists
-#         """
-# 
-#         index_mapping = {
-#             'properties': elasticsearch_schema,
-#         }
-# 
-#         if _meta:
-#             logging.info('==> index _meta: ' + pformat(_meta))
-#             index_mapping['_meta'] = _meta
-# 
-#         if not self.es.indices.exists(index=index_name):
-#             body = {
-#                 'mappings': index_mapping,
-#                 'settings': {
-#                     'number_of_shards': num_shards,
-#                     'number_of_replicas': 0,
-#                     'index.mapping.total_fields.limit': 10000,
-#                     'index.refresh_interval': -1,
-#                     'index.codec': 'best_compression',  # halves disk usage, no difference in query times
-#                 },
-#             }
-# 
-#             logging.info(
-#                 'create_mapping - elasticsearch schema: \n'
-#                 + pformat(elasticsearch_schema)
-#             )
-#             logging.info('==> creating elasticsearch index {}'.format(index_name))
-# 
-#             self.es.indices.create(index=index_name, body=body)
-#         else:
-#             if create_only:
-#                 raise ValueError('Index {} already exists'.format(index_name))
-# 
-#             logging.info(
-#                 '==> updating elasticsearch index {}. New schema:\n{}'.format(
-#                     index_name, pformat(elasticsearch_schema)
-#                 )
-#             )
-# 
-#             self.es.indices.put_mapping(index=index_name, body=index_mapping)
-# 
-#     def route_index_to_temp_es_cluster(self, index_name):
-#         """
-#         Apply shard allocation filtering rules to route the given index to elasticsearch loading nodes
-#         """
-#         self._update_settings(
-#             index_name,
-#             {
-#                 'index.routing.allocation.require._name': LOADING_NODES_NAME,
-#                 'index.routing.allocation.exclude._name': '',
-#             },
-#         )
-# 
-#     def route_index_off_temp_es_cluster(self, index_name):
-#         """
-#         Move any shards in the given index off of loading nodes
-#         """
-#         self._update_settings(
-#             index_name,
-#             {
-#                 'index.routing.allocation.require._name': '',
-#                 'index.routing.allocation.exclude._name': LOADING_NODES_NAME,
-#             },
-#         )
-# 
-#     def _update_settings(self, index_name, body):
-#         logging.info('==> Setting {} settings = {}'.format(index_name, body))
-# 
-#         self.es.indices.put_settings(index=index_name, body=body)
-# 
-#     def _get_index_meta(self, index_name):
-#         mappings = self.es.indices.get_mapping(index=index_name)
-#         return mappings.get(index_name, {}).get('mappings', {}).get('_meta', {})
-# 
-#     def wait_for_shard_transfer(self, index_name, num_attempts=1000):
-#         """
-#         Wait for shards to move off of the loading nodes before connecting to seqr
-#         """
-#         for _ in range(num_attempts):
-#             shards = self.es.cat.shards(index=index_name)
-#             if LOADING_NODES_NAME not in shards:
-#                 logging.warning('Shards are on {}'.format(shards))
-#                 return
-#             logging.warning(
-#                 'Waiting for {} shards to transfer off the es-data-loading nodes: \n{}'.format(
-#                     len(shards.strip().split('\n')), shards
-#                 )
-#             )
-#             time.sleep(5)
-# 
-#         raise Exception('Shards did not transfer off loading nodes')
-# 
-#     def export_table_to_elasticsearch(
-#         self,
-#         ht: hl.Table,
-#         index_name: str = 'data',
-#         block_size: int = 5000,
-#         num_shards: int = 10,
-#         delete_index_before_exporting: bool = True,
-#         elasticsearch_write_operation: str = ELASTICSEARCH_INDEX,
-#         ignore_elasticsearch_write_errors: bool = False,
-#         elasticsearch_mapping_id=None,
-#         field_name_to_elasticsearch_type_map=None,
-#         disable_doc_values_for_fields=(),
-#         disable_index_for_fields=(),
-#         field_names_replace_dot_with='_',
-#         func_to_run_after_index_exists=None,
-#         export_globals_to_index_meta=True,
-#         verbose=True,
-#         write_null_values=False,
-#         elasticsearch_config=None,
-#     ):
-#         """Create a new elasticsearch index to store the records in this table, and then export all records to it.
-# 
-#         Args:
-#             ht (Table): hail Table
-#             index_name (string): elasticsearch index name
-#             block_size (int): number of records to write in one bulk insert
-#             num_shards (int): number of shards to use for this index
-#                 (see https://www.elastic.co/guide/en/elasticsearch/guide/current/overallocation.html)
-#             delete_index_before_exporting (bool): Whether to drop and re-create the index before exporting.
-#             elasticsearch_write_operation (string): Can be one of these constants:
-#                     ELASTICSEARCH_INDEX
-#                     ELASTICSEARCH_CREATE
-#                     ELASTICSEARCH_UPDATE
-#                     ELASTICSEARCH_UPSERT
-#                 See https://www.elastic.co/guide/en/elasticsearch/hadoop/current/configuration.html#_operation
-#             ignore_elasticsearch_write_errors (bool): If True, elasticsearch errors will be logged, but won't cause
-#                 the bulk write call to throw an error. This is useful when, for example,
-#                 elasticsearch_write_operation='update', and the desired behavior is to update all documents that exist,
-#                 but to ignore errors for documents that don't exist.
-#             elasticsearch_mapping_id (str): if specified, sets the es.mapping.id which is the column name to use as the document ID
-#                 See https://www.elastic.co/guide/en/elasticsearch/hadoop/current/configuration.html#cfg-mapping
-#             field_name_to_elasticsearch_type_map (dict): (optional) a map of table field names to
-#                 their elasticsearch field spec - for example: {
-#                     'allele_freq': { 'type': 'half_float' },
-#                     ...
-#                 }.
-#                 See https://www.elastic.co/guide/en/elasticsearch/reference/current/mapping.html for
-#                 more details. Any values in this dictionary will override
-#                 the default type mapping derived from the hail table's row type.
-#                 Field names can be regular expressions.
-#             disable_doc_values_for_fields (tuple): (optional) list of field names (the way they will be
-#                 named in the elasticsearch index) for which to not store doc_values
-#                 (see https://www.elastic.co/guide/en/elasticsearch/reference/current/mapping-params.html)
-#             disable_index_for_fields (tuple): (optional) list of field names (the way they will be
-#                 named in the elasticsearch index) that shouldn't be indexed
-#                 (see https://www.elastic.co/guide/en/elasticsearch/reference/current/mapping-params.html)
-#             field_names_replace_dot_with (string): since '.' chars in field names are interpreted in
-#                 special ways by elasticsearch, set this arg to first go through and replace '.' with
-#                 this string in all field names. This replacement is not reversible (or atleast not
-#                 unambiguously in the general case) Set this to None to disable replacement, and fall back
-#                 on an encoding that's uglier, but reversible (eg. '.' will be converted to '_$dot$_')
-#             func_to_run_after_index_exists (function): optional function to run after creating the index, but before exporting any data.
-#             export_globals_to_index_meta (bool): whether to add table.globals object to the index _meta field:
-#                 (see https://www.elastic.co/guide/en/elasticsearch/reference/current/mapping-meta-field.html)
-#             verbose (bool): whether to print schema and stats
-#             write_null_values (bool): whether to write fields that are null to the index
-#             elasticsearch_config: The initial elasticsearch config from the caller
-#         """
-# 
-#         elasticsearch_config = elasticsearch_config or {}
-#         if (
-#             elasticsearch_write_operation is not None
-#             and elasticsearch_write_operation not in ELASTICSEARCH_WRITE_OPERATIONS
-#         ):
-#             raise ValueError(
-#                 'Unexpected value for elasticsearch_write_operation arg: '
-#                 + str(elasticsearch_write_operation)
-#             )
-# 
-#         if elasticsearch_write_operation is not None:
-#             elasticsearch_config['es.write.operation'] = elasticsearch_write_operation
-# 
-#         if (
-#             elasticsearch_write_operation
-#             in (ELASTICSEARCH_UPDATE, ELASTICSEARCH_UPSERT)
-#             or write_null_values
-#         ):
-#             # see https://www.elastic.co/guide/en/elasticsearch/hadoop/master/spark.html#spark-sql-write
-#             # 'By default, elasticsearch-hadoop will ignore null values in favor of not writing any field at all.
-#             # If updating/upserting, then existing field values may need to be overwritten with nulls
-#             elasticsearch_config['es.spark.dataframe.write.null'] = 'true'
-# 
-#         if elasticsearch_mapping_id is not None:
-#             elasticsearch_config['es.mapping.id'] = elasticsearch_mapping_id
-# 
-#         if ignore_elasticsearch_write_errors:
-#             # see docs in https://www.elastic.co/guide/en/elasticsearch/hadoop/current/errorhandlers.html
-#             elasticsearch_config['es.write.rest.error.handlers'] = 'log'
-#             elasticsearch_config[
-#                 'es.write.rest.error.handler.log.logging.name'
-#             ] = 'BulkErrors'
-# 
-#         if self._es_password:
-#             elasticsearch_config.update(
-#                 {
-#                     'es.net.http.auth.user': self._es_username,
-#                     'es.net.http.auth.pass': self._es_password,
-#                 }
-#             )
-# 
-#         if self._es_use_ssl:
-#             elasticsearch_config['es.net.ssl'] = 'true'
-#             # If using SSL, the instance is likely managed, in which case we
-#             # can't discover nodes.
-#             elasticsearch_config['es.nodes.wan.only'] = 'true'
-# 
-#         # encode any special chars in column names
-#         rename_dict = {}
-#         for field_name in ht.row_value.dtype.fields:
-#             encoded_name = field_name
-# 
-#             # optionally replace . with _ in a non-reversible way
-#             if field_names_replace_dot_with is not None:
-#                 encoded_name = encoded_name.replace('.', field_names_replace_dot_with)
-# 
-#             # replace all other special chars with an encoding that's uglier, but reversible
-#             encoded_name = encode_field_name(encoded_name)
-# 
-#             if encoded_name != field_name:
-#                 rename_dict[field_name] = encoded_name
-# 
-#         for original_name, encoded_name in rename_dict.items():
-#             logging.info('Encoding column name %s to %s', original_name, encoded_name)
-# 
-#         ht = ht.rename(rename_dict)
-# 
-#         if verbose:
-#             logging.info(pformat(ht.row_value.dtype))
-# 
-#         # create elasticsearch index with fields that match the ones in the table
-#         elasticsearch_schema = elasticsearch_schema_for_table(
-#             ht,
-#             disable_doc_values_for_fields=disable_doc_values_for_fields,
-#             disable_index_for_fields=disable_index_for_fields,
-#         )
-# 
-#         # override elasticsearch types
-#         if field_name_to_elasticsearch_type_map is not None:
-#             modified_elasticsearch_schema = dict(elasticsearch_schema)  # make a copy
-#             for (
-#                 field_name_regexp,
-#                 elasticsearch_field_spec,
-#             ) in field_name_to_elasticsearch_type_map.items():
-#                 match_count = 0
-#                 for key in elasticsearch_schema.keys():
-#                     if re.match(field_name_regexp, key):
-#                         modified_elasticsearch_schema[key] = elasticsearch_field_spec
-#                         match_count += 1
-# 
-#                 logging.info(f'{match_count} columns matched "{field_name_regexp}"')
-# 
-#             elasticsearch_schema = modified_elasticsearch_schema
-# 
-#         # optionally delete the index before creating it
-#         if delete_index_before_exporting and self.es.indices.exists(index=index_name):
-#             self.es.indices.delete(index=index_name)
-# 
-#         _meta = None
-#         if export_globals_to_index_meta:
-#             _meta = struct_to_dict(hl.eval(ht.globals))
-# 
-#         self.create_or_update_mapping(
-#             index_name, elasticsearch_schema, num_shards=num_shards, _meta=_meta
-#         )
-# 
-#         if func_to_run_after_index_exists:
-#             func_to_run_after_index_exists()
-# 
-#         logging.info(
-#             '==> exporting data to elasticsearch. Write mode: %s, blocksize: %d',
-#             elasticsearch_write_operation,
-#             block_size,
-#         )
-# 
-#         hl.export_elasticsearch(
-#             ht,
-#             self._host,
-#             int(self._port),
-#             index_name,
-#             '',
-#             block_size,
-#             elasticsearch_config,
-#             verbose,
-#         )
-# 
-#         """
-#         Potentially useful config settings for export_elasticsearch(..)
-#         (https://www.elastic.co/guide/en/elasticsearch/hadoop/current/configuration.html)
-# 
-#         es.write.operation // default: index (create, update, upsert)
-#         es.http.timeout // default 1m
-#         es.http.retries // default 3
-#         es.batch.size.bytes  // default 1mb
-#         es.batch.size.entries  // default 1000
-#         es.batch.write.refresh // default true  (Whether to invoke an index refresh or not after a bulk update has been completed)
-#         """
-# 
-#         self.es.indices.forcemerge(index=index_name, request_timeout=60)
-
-
-if __name__ == '__main__':
-    main()  # pylint: disable=E1120
