+++ test/test_check_pedigree.py
@@ -9,7 +9,7 @@ from cpg_utils import to_path, Path
 from cpg_utils.config import set_config_paths, update_dict
 from cpg_utils.workflows.utils import timestamp
 
-from jobs.python_scripts import check_pedigree
+from python_scripts import check_pedigree
 
 DEFAULT_CONF = """
 [workflow]
